# DQP enhacement and bug fixes