--- a/examples/index.html
+++ b/examples/index.html
@@ -328,9 +328,7 @@
                 </div>
                 <div class="col">
                     <h5>From Master</h5>
-                    <div class="video-container">
-                        <video id="videoPlayerFromMaster" class="remote-view" autoplay playsinline controls></video>
-                    </div>
+                    <div class="video-container"><video class="remote-view" autoplay playsinline controls></video></div>
                 </div>
             </div>
             <div class="row datachannel">